python/requirements.txt
black==24.10.0
click==8.1.8
flake8==7.1.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
typer==0.15.1
typing_extensions==4.12.2
